graded group
