+++ packages/perseus/src/widget-ai-utils/graded-group/graded-group-ai-utils.test.ts
@@ -5,7 +5,11 @@ import {testDependencies} from "../../../../../testing/test-dependencies";
 import * as Dependencies from "../../dependencies";
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./graded-group-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
+import type {CategorizerPromptJSON} from "../categorizer/categorizer-ai-utils";
+import type {ImagePromptJSON} from "../image/prompt-utils";
 import type {UserEvent} from "@testing-library/user-event";
 
 const question: PerseusRenderer = {
@@ -75,7 +79,7 @@ const question: PerseusRenderer = {
     },
 };
 
-describe("graded-group widget", () => {
+describe("GradedGroup AI utils", () => {
     let userEvent: UserEvent;
 
     beforeEach(() => {
@@ -88,6 +92,108 @@ describe("graded-group widget", () => {
         );
     });
 
+    it("it returns JSON with the expected format and fields when rendererJSON is undefined", () => {
+        const title = "title";
+        const rendererJSON = undefined;
+        const hintRendererJSON = {
+            content: "hint content",
+            widgets: {},
+        };
+        const result = getPromptJSON(title, rendererJSON, hintRendererJSON);
+
+        expect(result).toEqual({
+            type: "graded-group",
+            title,
+            content: "",
+            widgets: {},
+            hint: {
+                content: "hint content",
+                widgets: {},
+            },
+        });
+    });
+
+    it("it returns JSON with the expected format and fields", () => {
+        const title = "title";
+        const rendererJSON = {
+            title: "Metabolic strategies of bacteria",
+            content:
+                "1. **Which of the following statements about metabolic strategies of bacteria are true?**\n\n [[☃ categorizer 1]]",
+            images: {},
+            widgets: {
+                "categorizer 1": {
+                    type: "categorizer",
+                    options: {
+                        items: [
+                            "Some bacteria conduct photosynthesis and produce oxygen, much like plants.",
+                            "Bacteria are always autotrophic but they may get energy from either light or chemical sources.",
+                            "Some chemosynthetic bacteria introduce energy and fixed carbon into communities where photosynthesis is not possible (e.g., deep-sea vents).",
+                            "Some bacteria live symbiotically inside of host organisms and provide the host with nutrients.",
+                        ],
+                        categories: ["True", "False"],
+                    },
+                    userInput: {
+                        itemToCategoryMapping: [0, 0, 0, 0],
+                    },
+                } satisfies CategorizerPromptJSON,
+            },
+        };
+
+        const hintRendererJSON = {
+            content: "hint content",
+            widgets: {
+                "image 1": {
+                    type: "image",
+                    options: {
+                        altText: "alt text",
+                        title: "title",
+                        caption: "caption",
+                        imageUrl: "url",
+                    },
+                } satisfies ImagePromptJSON,
+            },
+        };
+        const result = getPromptJSON(title, rendererJSON, hintRendererJSON);
+
+        expect(result).toEqual({
+            type: "graded-group",
+            title,
+            content:
+                "1. **Which of the following statements about metabolic strategies of bacteria are true?**\n\n [[☃ categorizer 1]]",
+            images: {},
+            widgets: {
+                "categorizer 1": {
+                    type: "categorizer",
+                    options: {
+                        items: [
+                            "Some bacteria conduct photosynthesis and produce oxygen, much like plants.",
+                            "Bacteria are always autotrophic but they may get energy from either light or chemical sources.",
+                            "Some chemosynthetic bacteria introduce energy and fixed carbon into communities where photosynthesis is not possible (e.g., deep-sea vents).",
+                            "Some bacteria live symbiotically inside of host organisms and provide the host with nutrients.",
+                        ],
+                        categories: ["True", "False"],
+                    },
+                    userInput: {
+                        itemToCategoryMapping: [0, 0, 0, 0],
+                    },
+                },
+            },
+            hint: {
+                content: "hint content",
+                widgets: {
+                    "image 1": {
+                        type: "image",
+                        options: {
+                            altText: "alt text",
+                            title: "title",
+                            caption: "caption",
+                            imageUrl: "url",
+                        },
+                    },
+                },
+            },
+        });
+    });
     it("should get prompt json which matches the state of the UI when the hint is collapsed", async () => {
         // Arrange
         const {renderer} = renderQuestion(question);
